move axelar_gateway and executable out of interfaces
also move tests for types.rs
