+++ contracts/axelar-gateway/src/lib.rs
@@ -5,6 +5,8 @@ mod event;
 mod storage_types;
 
 pub mod contract;
+pub mod types;
+pub mod executable;
 
 #[cfg(all(target_family = "wasm", feature = "testutils"))]
 compile_error!("'testutils' feature is not supported on 'wasm' target");
@@ -17,3 +19,7 @@ mod test_gateway;
 
 #[cfg(test)]
 mod test_auth;
+
+#[cfg(test)]
+#[macro_use]
+extern crate std;
