+++ contracts/axelar-soroban-interfaces/src/lib.rs
@@ -1,8 +1,4 @@
 #![no_std]
 
-pub mod types;
-
-pub mod axelar_executable;
 pub mod axelar_gas_service;
-pub mod axelar_gateway;
 pub mod axelar_operators;
