--- contracts/axelar-soroban-interfaces/tests/testdata/messages_approval_hash.golden
@@ -1,1 +1,0 @@
-49f6a85aec4b4f72c667f2ba7950a692a59f462007abdbce0181e2982c0b602e
