--- contracts/axelar-soroban-interfaces/tests/testdata/weighted_signers_hash.golden
@@ -1,4 +1,0 @@
-[
-  "18b1ff823e202dab87cada621717e5be4955734bb973151eb489e6f1576ce3d4",
-  "4ad8f3015146ac68334fd405f90e6ca75fbf2c276b333a8747c9ba83d9c3f1f6"
-]
